refactor: change types directory location
