+++ packages/vibrant-core/src/index.ts
@@ -1,6 +1,6 @@
 export { Box } from './lib/Box';
 export type { BoxProps } from './lib/Box';
-export type { ReactElementChild, ResponsiveValue } from './lib/types';
+export type { ReactElementChild, ResponsiveValue } from './types';
 export type {
   BackgroundProps,
   BorderProps,
