+++ packages/vibrant-theme/src/index.ts
@@ -16,5 +16,5 @@ export type {
   TypographyKind,
   TypographyWeight,
   OpacityToken,
-} from './lib/types';
-export { baseColorTokens, onColorTokens, colorTokens } from './lib/types';
+} from './types';
+export { baseColorTokens, onColorTokens, colorTokens } from './types';
