+++ packages/vibrant-theme/src/types/Gradient.ts
